pyproject.toml
[project]

name = "ddpm_memo"
version = "0.0.1"
requires-python = ">= 3.9"
dependencies = [
  "diffusers",
  "torch",
  "pyyaml",
]

[build-system]

requires = ["setuptools >= 61.0"]
